Player traits
We now have code to read player traits. All custom game options options currently defined in the top-level game variant data use cobb::bitnumber or cobb::bytenumber as appropriate.

A cobb::bitnumber and a cobb::bitbool can be initialized with a default value using assignment syntax in its declaration (i.e. cobb::bitnumber<whatever> = 5).

You can now use enum classes as the underlying type for cobb::bitnumber without breaking things.

You can now cast cobb::bitnumber to the underlying integer type.

+++ native/src/ReachVariantTool/ReachVariantTool.cpp
@@ -19,7 +19,7 @@ int main() {
    printf("Header desc: %S\n", variant->contentHeader.data.description);
    printf("Embedded name: %S\n", variant->multiplayer.variantHeader.title);
    printf("Embedded desc: %S\n", variant->multiplayer.variantHeader.description);
-   printf("Loadout camera time: %d\n", variant->multiplayer.options.respawn.loadoutCamTime);
+   printf("Loadout camera time: %d\n", (int)variant->multiplayer.options.respawn.loadoutCamTime);
 
 
    return 0;
